+++ app/(marketing)/marketing-page/page.tsx
@@ -45,7 +45,9 @@ export default async function MarketingPage() {
           </Link>
         </div>
         <div className="mb-3 mt-[1.5rem] max-w-[900px] w-full">
-          <VideoPlayer videoSrc="https://utfs.io/f/08b0a37f-afd7-4623-b5cc-e85184528fce-1f02.mp4" />
+          <VideoPlayer 
+            videoSrc="https://utfs.io/f/08b0a37f-afd7-4623-b5cc-e85184528fce-1f02.mp4"
+          />
         </div>
         <div className="flex flex-col min-h-screen max-w-[900px] items-center mb-[2rem]">
           <article className="w-full mx-auto pb-8">